+++ configbox/chef/site-cookbooks/scriptbox/CHANGELOG.md
@@ -1,11 +1,11 @@
-wheezybox CHANGELOG
+scriptbox CHANGELOG
 ===================
 
-This file is used to list changes made in each version of the wheezybox cookbook.
+This file is used to list changes made in each version of the scriptbox cookbook.
 
 0.1.0
 -----
-- [your_name] - Initial release of wheezybox
+- [your_name] - Initial release of scriptbox
 
 - - -
 Check the [Markdown Syntax Guide](http://daringfireball.net/projects/markdown/syntax) for help with Markdown.